pyproject.toml
[project]
name = "taskforce"
version = "0.1.0"
description = "Taskforce HQ | Your intelligent AI-powered task management HQ."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=25.1",
    "django-stubs>=5.2.0",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "rest-framework-simplejwt>=0.0.2",
    "whitenoise>=6.9.0",
]

[tool.uv]
dev-dependencies = ["black>=24.1.0", "isort>=5.13.0", "ruff>=0.3.0"]
